requirements.txt
streamlit
google-cloud-storage
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu